Trait pallet_permissions::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Scope: Member + Parameter + MaxEncodedLen;
    type Role: Member + Parameter;
    type Storage: Member + Parameter + Properties<Property = Self::Role> + Default + MaxEncodedLen;
    type Editors: Contains<(Self::AccountId, Option<Self::Role>, Self::Scope, Self::Role)>;
    type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type MaxRolesPerScope: Get<u32>;
    type WeightInfo: WeightInfo;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Scope: Member + Parameter + MaxEncodedLen

type Role: Member + Parameter

type Storage: Member + Parameter + Properties<Property = Self::Role> + Default + MaxEncodedLen

type Editors: Contains<(Self::AccountId, Option<Self::Role>, Self::Scope, Self::Role)>

type AdminOrigin: EnsureOrigin<Self::RuntimeOrigin>

type MaxRolesPerScope: Get<u32>

type WeightInfo: WeightInfo